Use singularity for local driver for parallelization
Former-commit-id: 144b098b3f3c69467cf9a7c633f7b04628f23ded

+++ pqueens/schedulers/local_scheduler.py
@@ -1,6 +1,8 @@
 """ There should be a docstring """
 
+import pdb
 import os
+import sys
 from .scheduler import Scheduler
 
 
@@ -43,10 +45,27 @@ class LocalScheduler(Scheduler):
         Returns:
             bool: indicator if job is still alive
         """
-        try:
-            # Send an alive signal to proc
-            os.kill(process_id, 0)
-        except OSError:
+        alive = False
+        pdb.set_trace()
+        command_list = ['ps h -p', str(process_id)]
+        command_string = ' '.join(command_list)
+        stdout, _, p = super().run_subprocess(command_string)
+
+        if stdout:
+            sys.stderr.write("Job %d waiting in queue.\n" % (process_id))
+            alive = True
+        else:
+            sys.stderr.write("Job %d is held or suspended.\n" % (process_id))
+            alive = False
+
+        if not alive:
+            try:
+                # try to kill the job.
+                os.kill(process_id, 0)
+                sys.stderr.write("Killed job %d.\n" % (process_id))
+            except ValueError:
+                sys.stderr.write("Failed to kill job %d.\n" % (process_id))
+
             return False
         else:
             return True
